+++ iosDeKasegu/MethodEnum.swift
@@ -1,0 +1,16 @@
+//
+//  MethodEnums.swift
+//  iosDeKasegu
+//
+//  Created by 松栄健太 on 2019/05/25.
+//  Copyright © 2019 shonanhiratsuka. All rights reserved.
+//
+
+import Foundation
+
+enum MethodType {
+    case Siritori
+    case Mandara
+    case SixHats
+    case Osborne
+}
