+++ iosDeKasegu/Tutorial/SiritoriGuideViewController.swift
@@ -1,16 +1,14 @@
 //
-//  MandaraGuideViewController.swift
-//  iosDeKasegu
+//  SiritoriGuideViewController.swift
+//  
 //
-//  Created by 高木広大 on 2018/07/16.
-//  Copyright © 2018年 shonanhiratsuka. All rights reserved.
+//  Created by 高木広大 on 2018/07/12.
 //
 
 import UIKit
-//import PlaceholderTextView
 import AVFoundation
 
-class MandaraGuideViewController: BaseViewController,UIScrollViewDelegate {
+class SiritoriGuideViewController: BaseViewController,UIScrollViewDelegate {
     
     //UIPageControlのインスタンス作成
     let PageControl = UIPageControl()
@@ -23,7 +21,7 @@ class MandaraGuideViewController: BaseViewController,UIScrollViewDelegate {
         let Datum_yOfScrollView:CGFloat = UIScreen.main.bounds.height*0.1
         let WidthOfScrollView:CGFloat = UIScreen.main.bounds.width*0.9
         let HeightOfScrollView:CGFloat = UIScreen.main.bounds.height*0.9
-        let NumberOfPages:Int = 6
+        let NumberOfPages:Int = 5
         
         //スクロールビューの設定
         let scrollView = UIScrollView()
@@ -33,13 +31,13 @@ class MandaraGuideViewController: BaseViewController,UIScrollViewDelegate {
         scrollView.isPagingEnabled = true
         scrollView.delegate = self
         
-        // マンダラチャートの説明
+        // しりとり法の説明
         let IntroductionUIView = UIView(frame: CGRect(x: 0, y: 0, width: WidthOfScrollView, height: HeightOfScrollView))
         
         let TitelofIntroduction = UILabel()
         TitelofIntroduction.frame.size = CGSize(width: WidthOfScrollView, height: 0)
         TitelofIntroduction.center = CGPoint(x:self.view.center.x, y: HeightOfScrollView * 0.05)
-        let TitleText = "マンダラチャートとは"
+        let TitleText = "しりとり法とは"
         // attributedTextを作成する.
         let attributedText = NSMutableAttributedString(string: TitleText)
         let range = NSMakeRange(0, TitleText.count)
@@ -54,14 +52,13 @@ class MandaraGuideViewController: BaseViewController,UIScrollViewDelegate {
         
         let DetailofIntroduction = UILabel()
         DetailofIntroduction.frame = CGRect(x:0, y: TitelofIntroduction.frame.origin.y + TitelofIntroduction.frame.height + 15 ,width: WidthOfScrollView, height: 0)
-        DetailofIntroduction.text = "マンダラチャートは \n アイデアの拡大に使用するツールです。 \n 具体的には  \n 1.あるアイデアから9つの単語を連想 \n 2.連想した単語に関連する単語を9つ考える  \n という手順で進めます。\n 下は「りんごを育てるアプリ」の例です"
-        DetailofIntroduction.textAlignment = NSTextAlignment.left
+        DetailofIntroduction.text = "しりとり法は \n アイデアの発想を使用するツールです。 \n 具体的には \n 1.アイデアを出すテーマを決定 \n 2.最初のキーワードを決定 \n 3.しりとり形式で新しいキーワードを決定 \n 4.3のキーワードを含むアイデアを考案 \n 5.3と4を繰り返す \n という手順で進めます。 \n 下は「新しいアプリ」をテーマにした例です"
         //上記の文章の段数と合わせる
-        DetailofIntroduction.numberOfLines = 7
+        DetailofIntroduction.numberOfLines = 11
         DetailofIntroduction.font = UIFont.systemFont(ofSize: 17)
         DetailofIntroduction.sizeToFit()
         
-        let ImageofIntroduction = UIImage(named: "MandaraExample.png")
+        let ImageofIntroduction = UIImage(named: "SiritoriExample.png")
         let UIImageofIntroduction = UIImageView(image: ImageofIntroduction)
         UIImageofIntroduction.frame.size.width = WidthOfScrollView
         //元の画像のアスペクト比を保ったまま、画像を縮小
@@ -69,8 +66,8 @@ class MandaraGuideViewController: BaseViewController,UIScrollViewDelegate {
         //縮小した後の画像のサイズと座標を取得
         let FrameOfImage = AVMakeRect(aspectRatio: (ImageofIntroduction?.size)!, insideRect: UIImageofIntroduction.bounds)
         //縮小した後の画像の座標に合わせてUIViewもオフセット
-        UIImageofIntroduction.frame.origin = CGPoint(x: 0, y:  (DetailofIntroduction.frame.origin.y - FrameOfImage.origin.y) + DetailofIntroduction.frame.height + 15)
-
+        UIImageofIntroduction.frame.origin = CGPoint(x: 0, y:  (DetailofIntroduction.frame.origin.y - FrameOfImage.origin.y) + DetailofIntroduction.frame.height + 10)
+        
         let DetailofApp = UILabel()
         DetailofApp.frame = CGRect(x:0, y: UIImageofIntroduction.frame.origin.y + (UIImageofIntroduction.frame.height - FrameOfImage.origin.y) + 15 ,width: WidthOfScrollView, height: 0 )
         DetailofApp.text = "左にスワイプするとアプリの使い方を \n 見ることが出来ます。 \n 途中の状態で保存して再開することも　\n 出来ます。（保存は自動で行われます。）"
@@ -86,7 +83,7 @@ class MandaraGuideViewController: BaseViewController,UIScrollViewDelegate {
         
         // 1枚目の画像
         let FirstUIView = UIView(frame: CGRect(x: WidthOfScrollView, y: 0, width: WidthOfScrollView, height: HeightOfScrollView))
-        let FirstImage = UIImageView(image: UIImage(named: "MandaraTutorialImage1.png"))
+        let FirstImage = UIImageView(image: UIImage(named: "SiritoriTutorialImage1.png"))
         FirstImage.frame.size.height = HeightOfScrollView * 0.8
         FirstImage.center = CGPoint(x: WidthOfScrollView * 0.5, y: HeightOfScrollView * 0.4)
         FirstImage.contentMode = UIView.ContentMode.scaleAspectFit
@@ -104,14 +101,14 @@ class MandaraGuideViewController: BaseViewController,UIScrollViewDelegate {
         
         // 2枚目の画像
         let SecondUIView = UIView(frame: CGRect(x: WidthOfScrollView * 2, y: 0, width: WidthOfScrollView, height: HeightOfScrollView))
-        let SecondImage = UIImageView(image: UIImage(named: "MandaraTutorialImage2.png"))
+        let SecondImage = UIImageView(image: UIImage(named: "SiritoriTutorialImage2.png"))
         SecondImage.frame.size.height = HeightOfScrollView * 0.8
         SecondImage.center = CGPoint(x: WidthOfScrollView * 0.5, y: HeightOfScrollView * 0.4)
         SecondImage.contentMode = UIView.ContentMode.scaleAspectFit
         
         let SecondLable = UILabel()
-        SecondLable.text = "ピンチアウトする事で \n 編集する箇所を拡大できます"
-        SecondLable.numberOfLines = 2
+        SecondLable.text = "最初のフレーズを決定 \n （デフォルト:しりとり）\n 最初のフレーズとしりとりをして \n キーワード1を決定"
+        SecondLable.numberOfLines = 4
         SecondLable.font = UIFont.systemFont(ofSize: 20.0)
         SecondLable.frame = CGRect(x:0, y: HeightOfScrollView * 0.72,width: WidthOfScrollView, height: HeightOfScrollView * 0.3)
         SecondLable.adjustsFontSizeToFitWidth = true
@@ -122,13 +119,13 @@ class MandaraGuideViewController: BaseViewController,UIScrollViewDelegate {
         
         // 3枚目の画像
         let ThirdUIView = UIView(frame: CGRect(x: WidthOfScrollView * 3, y: 0, width: WidthOfScrollView, height: HeightOfScrollView))
-        let ThirdImage = UIImageView(image: UIImage(named: "MandaraTutorialImage3.png"))
+        let ThirdImage = UIImageView(image: UIImage(named: "SiritoriTutorialImage3.png"))
         ThirdImage.frame.size.height = HeightOfScrollView * 0.8
         ThirdImage.center = CGPoint(x: WidthOfScrollView * 0.5, y: HeightOfScrollView * 0.4)
         ThirdImage.contentMode = UIView.ContentMode.scaleAspectFit
         
         let ThirdLable = UILabel()
-        ThirdLable.text = "先ずは赤マスの隣の緑マスに \n テーマから連想される \n 単語を入力しましょう \n 外側には自動コピーされます"
+        ThirdLable.text = "キーワード1を含むアイデア1を考案 \n 次へボタンをタップして \n キーワード2のカードを作成 \n アイデア数は無制限です"
         ThirdLable.numberOfLines = 4
         ThirdLable.font = UIFont.systemFont(ofSize: 20.0)
         ThirdLable.frame = CGRect(x:0, y: HeightOfScrollView * 0.72,width: WidthOfScrollView, height: HeightOfScrollView * 0.3)
@@ -138,39 +135,21 @@ class MandaraGuideViewController: BaseViewController,UIScrollViewDelegate {
         ThirdUIView.addSubview(ThirdLable)
         scrollView.addSubview(ThirdUIView)
         
-        // 4枚目の画像
-        let FourthUIView = UIView(frame: CGRect(x: WidthOfScrollView * 4, y: 0, width: WidthOfScrollView, height: HeightOfScrollView))
-        let FourthImage = UIImageView(image: UIImage(named: "MandaraTutorialImage4.png"))
-        FourthImage.frame.size.height = HeightOfScrollView * 0.8
-        FourthImage.center = CGPoint(x: WidthOfScrollView * 0.5, y: HeightOfScrollView * 0.4)
-        FourthImage.contentMode = UIView.ContentMode.scaleAspectFit
-        
-        let FourthLable = UILabel()
-        FourthLable.text = "外側の緑マスの隣の青マスに \n 緑マスの単語に関連する \n 単語を入力しましょう"
-        FourthLable.numberOfLines = 3
-        FourthLable.font = UIFont.systemFont(ofSize: 20.0)
-        FourthLable.frame = CGRect(x:0, y: HeightOfScrollView * 0.72,width: WidthOfScrollView, height: HeightOfScrollView * 0.3)
-        FourthLable.adjustsFontSizeToFitWidth = true
-        FourthLable.minimumScaleFactor = 0.3
-        FourthUIView.addSubview(FourthImage)
-        FourthUIView.addSubview(FourthLable)
-        scrollView.addSubview(FourthUIView)
-        
-        //マンダラテーマ画面へ遷移するボタン
-        let ButtonUIView = UIView(frame: CGRect(x: WidthOfScrollView * 5, y: 0, width: WidthOfScrollView, height: HeightOfScrollView))
-        let MandaraButton = UIButton()
-        MandaraButton.frame.size = CGSize(width: WidthOfScrollView, height: HeightOfScrollView * 0.3)
-        MandaraButton.center = CGPoint(x:WidthOfScrollView * 0.5, y:HeightOfScrollView * 0.5)
-        MandaraButton.backgroundColor = UIColor(hex: "59ABE3")
-        MandaraButton.addTarget(self, action: #selector(self.movetotheme(sender:)), for: .touchUpInside)
+        //しりとりテーマ画面へ遷移するボタン
+        let ButtonUIView = UIView(frame: CGRect(x: WidthOfScrollView * 4, y: 0, width: WidthOfScrollView, height: HeightOfScrollView))
+        let SiritoriButton = UIButton()
+        SiritoriButton.frame.size = CGSize(width: WidthOfScrollView, height: HeightOfScrollView * 0.3)
+        SiritoriButton.center = CGPoint(x:WidthOfScrollView * 0.5, y:HeightOfScrollView * 0.5)
+        SiritoriButton.backgroundColor = UIColor(hex: "F4D03F")
+        SiritoriButton.addTarget(self, action: #selector(self.movetotheme(sender:)), for: .touchUpInside)
         let ButtonLable = UILabel()
-        ButtonLable.text = "マンダラチャートを使う \n （テーマ作成画面に移動します）"
+        ButtonLable.text = "しりとり法を使う \n （テーマ作成画面に移動します）"
         ButtonLable.numberOfLines = 2
         ButtonLable.textAlignment = NSTextAlignment.center
         ButtonLable.font = UIFont.systemFont(ofSize: 20)
-        ButtonLable.frame = CGRect(x:0, y:0,width: MandaraButton.frame.width, height: MandaraButton.frame.height)
-        MandaraButton.addSubview(ButtonLable)
-        ButtonUIView.addSubview(MandaraButton)
+        ButtonLable.frame = CGRect(x:0, y:0,width: SiritoriButton.frame.width, height: SiritoriButton.frame.height)
+        SiritoriButton.addSubview(ButtonLable)
+        ButtonUIView.addSubview(SiritoriButton)
         scrollView.addSubview(ButtonUIView)
         
         // スクロールビューを追加
@@ -197,9 +176,11 @@ class MandaraGuideViewController: BaseViewController,UIScrollViewDelegate {
     }
     
     @objc func movetotheme(sender: Any) {
-        self.performSegue(withIdentifier: "toMandaraTheme", sender: nil)
+        let storyboard = UIStoryboard(name: "Siritori", bundle: nil)
+        let destinationNC = storyboard.instantiateViewController(withIdentifier: "SiritoriThemesNC") as! UINavigationController
+        self.navigationController?.pushViewController(destinationNC.topViewController!, animated: true)
     }
-    
+
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
